Rename networks, remove "versatile"
The versatile name refers to the pretrained model
that has been trained on a large variety of data.
But any model can be used with the networks. Thus
they are now renamed "fluo" and "he" in refeference
to the type of data they target.
