+++ crates/cellcast/src/networks/stardist/fluo_2d.rs
