+++ crates/cellcast/src/networks/stardist/he_2d.rs
