+++ crates/cellcast/src/networks/stardist/mod.rs
@@ -1,3 +1,3 @@
 pub mod demo_3d;
-pub mod versatile_fluo_2d;
-pub mod versatile_he_2d;
+pub mod fluo_2d;
+pub mod he_2d;
